Move files a bit.
